refactor: 패키지 구조를 레이어드 계층 기반으로 수정
