--- src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,54 +1,0 @@
-package roomescape.reservation.controller;
-
-import java.net.URI;
-import java.time.LocalDate;
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.controller.dto.ReservationCreateRequest;
-import roomescape.reservation.controller.dto.ReservationResponse;
-import roomescape.reservation.service.ReservationService;
-
-@RestController
-@RequestMapping("/reservations")
-public class ReservationController {
-
-    private final ReservationService reservationService;
-
-    public ReservationController(final ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ReservationResponse>> getReservations() {
-        List<ReservationResponse> reservations = reservationService.getAll().stream()
-                .map(ReservationResponse::from)
-                .toList();
-        return ResponseEntity.ok().body(reservations);
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationCreateRequest reservationRequest) {
-        ReservationResponse reservation = ReservationResponse.from(reservationService.save(
-                reservationRequest.name(),
-                reservationRequest.date(),
-                reservationRequest.themeId(),
-                reservationRequest.timeId()
-        ));
-
-        return ResponseEntity.created(URI.create("/reservations/" + reservation.id()))
-                .body(reservation);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
-}
