--- src/main/java/roomescape/reservation/controller/dto/ReservationCreateRequest.java
@@ -1,16 +1,0 @@
-package roomescape.reservation.controller.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalDate;
-
-public record ReservationCreateRequest(
-        String name,
-
-        @JsonFormat(pattern = "yyyy-MM-dd")
-        LocalDate date,
-
-        Long themeId,
-
-        Long timeId
-) {
-}
