--- src/main/java/roomescape/reservationtime/controller/dto/ReservationTimeCreateRequest.java
@@ -1,10 +1,0 @@
-package roomescape.reservationtime.controller.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalTime;
-
-public record ReservationTimeCreateRequest(
-        @JsonFormat(pattern = "HH:mm")
-        LocalTime startAt
-) {
-}
