--- src/main/java/roomescape/reservationtime/controller/dto/ReservationTimeResponse.java
@@ -1,17 +1,0 @@
-package roomescape.reservationtime.controller.dto;
-
-import java.time.LocalTime;
-import roomescape.reservationtime.domain.ReservationTime;
-
-public record ReservationTimeResponse(
-        Long id,
-        LocalTime startAt
-) {
-
-    public static ReservationTimeResponse from(final ReservationTime reservationTime) {
-        return new ReservationTimeResponse(
-                reservationTime.getId(),
-                reservationTime.getStartAt()
-        );
-    }
-}
