--- src/main/java/roomescape/reservationtime/repository/ReservationTimeRepository.java
@@ -1,20 +1,0 @@
-package roomescape.reservationtime.repository;
-
-import java.time.LocalTime;
-import java.util.List;
-import java.util.Optional;
-import roomescape.reservationtime.domain.ReservationTime;
-
-public interface ReservationTimeRepository {
-
-    List<ReservationTime> findAll();
-
-    Optional<ReservationTime> findById(long timeId);
-
-    void deleteById(long timeId);
-
-    ReservationTime save(ReservationTime reservationTime);
-
-    boolean existsByStartAt(LocalTime startAt);
-
-}
