--- src/main/java/roomescape/theme/controller/dto/ThemeCreateRequest.java
@@ -1,8 +1,0 @@
-package roomescape.theme.controller.dto;
-
-public record ThemeCreateRequest(
-        String name,
-        String description,
-        String thumbnailUrl
-) {
-}
