--- src/main/java/roomescape/theme/repository/ThemeRepository.java
@@ -1,21 +1,0 @@
-package roomescape.theme.repository;
-
-import java.util.List;
-import java.util.Optional;
-import roomescape.theme.domain.Theme;
-
-public interface ThemeRepository {
-
-    List<Theme> findAll();
-
-    Optional<Theme> findById(long id);
-
-    void deleteById(long id);
-
-    Theme save(Theme theme);
-
-    boolean existsByName(String name);
-
-    List<Theme> findPopularThemes(int period, int limit);
-
-}
