+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.domain.reservationtime.ReservationTime;
 
 class ReservationTimeTest {
