+++ src/test/java/roomescape/domain/ThemeTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.theme.domain.Theme;
+import roomescape.domain.theme.Theme;
 
 class ThemeTest {
